requirements.txt
google-generativeai==0.3.2
requests==2.32.4
python-dotenv==1.0.1
